# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OmniCLI Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /search:
    get:
      operationId: searchFiles
      tags: [search]
      summary: Full-text search across indexed files
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: types
          in: query
          schema:
            type: string
            description: Comma-separated content types (files,code,pdf,json,logs,sqlite,zip)
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: regex
          in: query
          schema:
            type: boolean
            default: false
        - name: caseSensitive
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /search/index:
    get:
      operationId: getSearchIndexInfo
      tags: [search]
      summary: Get search index information and stats
      responses:
        "200":
          description: Search index info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchIndexInfo"
    post:
      operationId: triggerSearchIndex
      tags: [search]
      summary: Trigger a search index rebuild for given paths
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchIndexRequest"
      responses:
        "200":
          description: Index triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchIndexInfo"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Files ──────────────────────────────────────────────────────────────────
...
  /workspace/notes:
    get:
      operationId: listNotes
      tags: [workspace]
      summary: List all notes
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [workspace]
      summary: Create a new note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workspace/notes/{id}:
    get:
      operationId: getNote
      tags: [workspace]
      summary: Get a note by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateNote
      tags: [workspace]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteNote
      tags: [workspace]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /workspace/todos:
    get:
      operationId: listTodos
      tags: [workspace]
      summary: List all todos
      parameters:
        - name: done
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Todo"
    post:
      operationId: createTodo
      tags: [workspace]
      summary: Create a todo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoInput"
      responses:
        "201":
          description: Created todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workspace/todos/{id}:
    patch:
      operationId: updateTodo
      tags: [workspace]
      summary: Update a todo (toggle done, edit description)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TodoUpdate"
      responses:
        "200":
          description: Updated todo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Todo"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTodo
      tags: [workspace]
      summary: Delete a todo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /workspace/snippets:
    get:
      operationId: listSnippets
      tags: [workspace]
      summary: List all code snippets
      parameters:
        - name: language
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Snippets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Snippet"
    post:
      operationId: createSnippet
      tags: [workspace]
      summary: Save a code snippet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SnippetInput"
      responses:
        "201":
          description: Created snippet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Snippet"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workspace/snippets/{id}:
    delete:
      operationId: deleteSnippet
      tags: [workspace]
      summary: Delete a snippet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Dashboard ──────────────────────────────────────────────────────────────
...
    SearchResult:
      type: object
      required: [path, fileType, sizeBytes, modified, snippet]
      properties:
        path:
          type: string
        fileType:
          type: string
        sizeBytes:
          type: integer
        modified:
          type: integer
        snippet:
          type: ["string", "null"]
        rank:
          type: ["number", "null"]
    SearchResponse:
      type: object
      required: [query, results, total, durationMs]
      properties:
        query:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
        total:
          type: integer
        durationMs:
          type: number
    SearchIndexInfo:
      type: object
      required: [totalEntries, lastUpdated, indexPath, status]
      properties:
        totalEntries:
          type: integer
        lastUpdated:
          type: ["string", "null"]
        indexPath:
          type: string
        status:
          type: string
          enum: [ready, indexing, empty, error]
        indexedPaths:
          type: array
          items:
            type: string
    SearchIndexRequest:
      type: object
      required: [paths]
      properties:
        paths:
          type: array
          items:
            type: string
          minItems: 1
        rebuild:
          type: boolean
          default: false

    # ── Files ──────────────────────────────────────────────────────────────────
...
    Note:
      type: object
      required: [id, title, body, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        body:
          type: string
        tags:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    NoteInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        body:
          type: string
        tags:
          type: string
    NoteUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        body:
          type: string
        tags:
          type: string
    Todo:
      type: object
      required: [id, description, done, createdAt]
      properties:
        id:
          type: integer
        description:
          type: string
        done:
          type: boolean
        dueAt:
          type: ["string", "null"]
        createdAt:
          type: string
    TodoInput:
      type: object
      required: [description]
      properties:
        description:
          type: string
          minLength: 1
        dueAt:
          type: string
    TodoUpdate:
      type: object
      properties:
        description:
          type: string
          minLength: 1
        done:
          type: boolean
        dueAt:
          type: ["string", "null"]
    Snippet:
      type: object
      required: [id, name, body, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        language:
          type: ["string", "null"]
        body:
          type: string
        createdAt:
          type: string
    SnippetInput:
      type: object
      required: [name, body]
      properties:
        name:
          type: string
          minLength: 1
        language:
          type: string
        body:
          type: string
          minLength: 1
...