# lib/api-zod/src/generated/types/snippet.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * OmniCLI Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Snippet {
  id: number;
  name: string;
  /** @nullable */
  language?: string | null;
  body: string;
  createdAt: string;
}